+++ trueseeing/app/cmd/android/show.py
@@ -1,21 +1,20 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, NamedTuple
 
 import re
-import sys
 from collections import deque
 
 from trueseeing.core.model.cmd import CommandMixin
-from trueseeing.core.ui import ui
+from trueseeing.core.ui import ui, OpFormatter
 
 if TYPE_CHECKING:
-  from typing import Optional, TypedDict
+  from typing import Optional
   from trueseeing.api import CommandHelper, Command, CommandMap
   from trueseeing.core.android.context import APKContext
 
-  class QualNameInfo(TypedDict):
-    path: str
-    sig: str
+class QualNameInfo(NamedTuple):
+  clazz: str
+  method: Optional[str]
 
 class ShowCommand(CommandMixin):
   def __init__(self, helper: CommandHelper) -> None:
@@ -27,32 +26,27 @@ class ShowCommand(CommandMixin):
 
   def get_commands(self) -> CommandMap:
     return {
-      'pd':dict(e=self._show_disasm, n='pd[!] qualname [output.smali]', d='show disassembled class/method'),
-      'pd!':dict(e=self._show_disasm),
+      'pd':dict(e=self._show_disasm, n='pd qualname', d='show disassembled class/method'),
       'pk':dict(e=self._show_solved_constant, n='pk 0xop index', d='guess and show what constant would flow into the index-th arg of op (!: try harder)'),
       'pt':dict(e=self._show_solved_typeset, n='pt 0xop index', d='guess and show what type would flow into the index-th arg of op'),
     }
 
-  async def _show_disasm(self, args: deque[str]) -> None:
-    outfn: Optional[str] = None
+  def _output_as_untagged_listing(self, is_header: bool, line: str) -> None:
+    if not is_header:
+      ui.info(line)
 
+  async def _show_disasm(self, args: deque[str]) -> None:
     self._helper.require_target()
 
-    cmd = args.popleft()
+    _ = args.popleft()
 
-    if not args:
+    try:
+      qn = args.popleft()
+    except IndexError:
       ui.fatal('need a qualname')
 
-    qn = args.popleft()
-
-    import os
-
-    if args:
-      outfn = args.popleft()
-      if os.path.exists(outfn) and not cmd.endswith('!'):
-        ui.fatal('outfile exists; force (!) to overwrite')
-
     context: APKContext = self._helper.get_context().require_type('apk')
+    await context.analyze()
 
     try:
       c = self._parse_qualname(qn)
@@ -60,40 +54,17 @@ class ShowCommand(CommandMixin):
       ui.fatal(f'invalid qualname (try quoting): {qn}')
 
     with context.store().query().scoped() as q:
-      for _, d in q.file_enum('smali%/{}'.format(c['path'])):
-        if outfn is None:
-          f = sys.stdout.buffer
-        else:
-          f = open(outfn, 'wb')
-        try:
-          if not c['sig']:
-            f.write(d)
-          else:
-            from io import BytesIO
-            pat = r'\.method.*? {}$'.format(re.escape(c['sig'])).encode('utf-8')
-            state = 0
-            for l in BytesIO(d):
-              if state == 0:
-                if re.match(pat, l):
-                  f.write(l)
-                  state = 1
-              elif state == 1:
-                f.write(l)
-                if l == b'.end method\n':
-                  state = 2
-              elif state == 2:
-                break
-        finally:
-          if f != sys.stdout.buffer:
-            f.close()
+      formatter = OpFormatter(q)
+      for is_header, line in formatter.format(q.body(c.clazz, c.method)):
+        self._output_as_untagged_listing(is_header, line)
 
   def _parse_qualname(self, n: str) -> QualNameInfo:
     m = re.fullmatch('(L[^ ]+?;)(->[^ ]+?)?', n)
     if m is None:
       raise ValueError('invalid dalvik name: {n}') # XXX
-    return dict(
-      path='{}.smali'.format(m.group(1)[1:-1]),
-      sig=m.group(2)[2:] if m.group(2) else '',
+    return QualNameInfo(
+      clazz=m.group(1),
+      method=m.group(2)[2:],
     )
 
   async def _show_solved_constant(self, args: deque[str]) -> None:
