+++ trueseeing/core/ui.py
@@ -8,10 +8,12 @@ from pubsub import pub
 from trueseeing.core.exc import FatalError
 
 if TYPE_CHECKING:
-  from typing import NoReturn, Optional, TextIO, Set, Any, Iterator
+  from typing import NoReturn, Optional, TextIO, Set, Any, Iterator, Iterable, Tuple
   from typing_extensions import Final
   from progressbar import ProgressBar
   from trueseeing.core.model.issue import Issue
+  from trueseeing.core.android.model import Op
+  from trueseeing.core.android.db import APKQuery
 
 class UI:
   DEBUG: Final = 0
@@ -398,5 +400,23 @@ class ScanProgressReporter:
   def _issue(self, issue: Issue) -> None:
     self._CN.note(issue)
 
+class OpFormatter:
+  def __init__(self, q: APKQuery, indent: int = 4) -> None:
+    self._q = q
+    self._indent = indent
+
+  def format(self, ops: Iterable[Op]) -> Iterator[Tuple[bool, str]]:
+    focus: Optional[str] = None
+    for op in ops:
+      qn = self._q.qualname_of(op.addr)
+      if qn is None:
+        qn = self._q.class_name_of(op.addr)
+      if qn != focus:
+        yield True, f'{qn}:'
+        focus = qn
+      yield False, '{ind}{op}'.format(
+        ind=' '*self._indent,
+        op='{id:08x}{sep}{l}'.format(sep=' '*4, id=op.addr, l=op.l.lstrip())
+      )
 
 ui = UI()
